Move to mining-pipeline module
