+++ modules/mining-pipeline/ccspan-pattern-detector/build.gradle
@@ -1,5 +1,5 @@
 dependencies {
-    compile project(":pipeline")
+    compile project(":mining-pipeline")
 
     testCompile group: "com.nhaarman", name: "mockito-kotlin", version: mockitoKotlinVersion
 }
