+++ modules/mining-pipeline/ccspan-pattern-detector/gradle.properties
