+++ modules/mining-pipeline/ccspan-pattern-detector/src/test/kotlin/org/cafejojo/schaapi/miningpipeline/patterndetector/ccspan/CCSpanTest.kt
@@ -1,4 +1,4 @@
-package org.cafejojo.schaapi.pipeline.patterndetector.ccspan
+package org.cafejojo.schaapi.miningpipeline.patterndetector.ccspan
 
 import org.assertj.core.api.Assertions.assertThat
 import org.cafejojo.schaapi.models.Node
