+++ modules/mining-pipeline/ccspan-pattern-detector/src/test/kotlin/org/cafejojo/schaapi/miningpipeline/patterndetector/ccspan/TestNodeComparator.kt
@@ -1,4 +1,4 @@
-package org.cafejojo.schaapi.pipeline.patterndetector.ccspan
+package org.cafejojo.schaapi.miningpipeline.patterndetector.ccspan
 
 import org.cafejojo.schaapi.models.GeneralizedNodeComparator
 import org.cafejojo.schaapi.models.Node
@@ -8,7 +8,7 @@ import org.cafejojo.schaapi.models.Node
  *
  * Two [Node]s are considered equal if they have the same reference.
  */
-class TestNodeComparator<N : Node> : GeneralizedNodeComparator<N> {
+internal class TestNodeComparator<N : Node> : GeneralizedNodeComparator<N> {
     override fun structuresAreEqual(template: N, instance: N) = template === instance
     override fun generalizedValuesAreEqual(template: N, instance: N) = template === instance
     override fun satisfies(template: N, instance: N) = template === instance
